+++ src/main/java/fr/esgi/grp9/uparserbackend/mail/service/IEmailService.java
@@ -1,7 +1,7 @@
-package fr.esgi.grp9.uparserbackend.mail.domain;
+package fr.esgi.grp9.uparserbackend.mail.service;
 
 import fr.esgi.grp9.uparserbackend.mail.domain.Email;
 
-public interface EmailService {
+public interface IEmailService {
     void sendResetPasswordEmail(Email email);
 }
